requirements.txt
pyTelegramBotAPI
google-genai
python-dotenv

